backend/pyproject.toml
[project]
name = "outbound-ai-calling-backend"
version = "0.1.0"
description = "FastAPI backend for the outbound AI calling system"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0",
  "sqlalchemy>=2.0.0,<3.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "PyJWT>=2.10.0,<3.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "httpx>=0.28.0,<1.0.0",
  "python-multipart>=0.0.12",
  "alembic>=1.14.0",
  "phonenumberslite>=8.13.0,<9.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
